🚚 rename: utils 함수 통일 (Jira Ticket : CCC-66)
