--- utils/dateFormat/index.ts
@@ -1,3 +1,0 @@
-export { default as formatToDate } from './FormatToDate';
-export { default as formatToTime } from './FormatToTime';
-export { default as formatTimeOrDate } from './FormatTimeOrDate';
